Make for loop counter type same as count's

+++ SafeCategories.h
@@ -102,7 +102,7 @@ static NSArray* get_related_classes(Class baseClass) {
   // Now get all subclasses of the base class
   unsigned int count = 0;
   Class* runtimeClasses = objc_copyClassList(&count);
-  for (NSInteger i = 0; i < count; i++) {
+  for (unsigned int i = 0; i < count; i++) {
     Class superClass = runtimeClasses[i];
     do {
       superClass = class_getSuperclass(superClass);
